# api/swagger.yaml
basePath: /
definitions:
...
  employees.AttachRestaurantRequest:
    properties:
      code:
        type: string
      position:
        type: string
    type: object
  employees.BatchRequest:
    properties:
      ids:
        items:
          type: string
        type: array
    type: object
  employees.ComponentView:
    properties:
      brand:
        type: string
      flavour:
        type: string
      percent:
        type: integer
    type: object
  employees.CreateRestaurantRequest:
    properties:
      name:
        example: Bar One
        type: string
    type: object
  employees.EmployeeResponse:
    properties:
      firstName:
        type: string
      id:
        type: string
      lastName:
        type: string
      middleName:
        type: string
      position:
        type: string
      shortName:
        type: string
    type: object
  employees.ListByCodeRequest:
    properties:
      code:
        type: string
    type: object
  employees.LoginResponse:
    properties:
      employees:
        items:
          $ref: '#/definitions/employees.EmployeeResponse'
        type: array
      restaurant:
        $ref: '#/definitions/employees.RestaurantBrief'
    type: object
  employees.RateEmployeeRequest:
    properties:
      score:
        example: 5
        type: integer
      userId:
        type: string
    type: object
  employees.RatingAgg:
    properties:
      average:
        type: number
      count:
        type: integer
    type: object
  employees.RatingItem:
    properties:
      score:
        type: integer
      updatedAt:
        type: string
      userId:
        type: string
    type: object
  employees.RecipeFeedbackItem:
    properties:
      components:
        items:
          $ref: '#/definitions/employees.ComponentView'
        type: array
      orderRecipeId:
        type: string
      recipeId:
        type: string
      recipeName:
        type: string
      review:
        type: string
      score:
        type: integer
      strength:
        type: integer
      updatedAt:
        type: string
    type: object
  employees.RegisterEmployeeRequest:
    properties:
      code:
        example: SECRET123
        type: string
      firstName:
        type: string
      lastName:
        type: string
      middleName:
        type: string
      position:
        example: Старший мастер
        type: string
      shortName:
        type: string
    type: object
  employees.RegisterEmployeeResponse:
    properties:
      employee:
        $ref: '#/definitions/employees.EmployeeResponse'
      restaurantId:
        type: string
    type: object
  employees.RestaurantBrief:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  employees.RestaurantResponse:
    properties:
      code:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  employees.SetShiftRequest:
    properties:
      code:
        type: string
      employeeIds:
        items:
          type: string
        type: array
    type: object
  employees.ShiftMaster:
    properties:
      firstName:
        type: string
      id:
        type: string
      lastName:
        type: string
      middleName:
        type: string
      position:
        type: string
      rating:
        type: number
      ratingCount:
        type: integer
      shortName:
        type: string
    type: object
  employees.UpdateEmployeeRequest:
    properties:
      firstName:
        type: string
      lastName:
        type: string
      middleName:
        type: string
      shortName:
        type: string
    type: object
...
  httpx.APIError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  httpx.Envelope:
    properties:
      data: {}
      error:
        $ref: '#/definitions/httpx.APIError'
    type: object
...
info:
  contact: {}
  description: Backend for the mixMaster restaurant ordering / recipes app.
  title: mixMaster API
  version: 0.1.0
paths:
  /employees:
    post:
      consumes:
      - application/json
      parameters:
      - description: Employee data + restaurant code
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/employees.RegisterEmployeeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/httpx.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/employees.RegisterEmployeeResponse'
              type: object
        "404":
          description: unknown code
          schema:
            $ref: '#/definitions/httpx.Envelope'
      summary: Register an employee at a restaurant (by code)
      tags:
      - employees
  /employees/{id}:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to change
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/employees.UpdateEmployeeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/employees.EmployeeResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.Envelope'
      summary: Update an employee (partial)
      tags:
      - employees
  /employees/{id}/rating:
    get:
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/employees.RatingAgg'
              type: object
      summary: Get a master's aggregate rating
      tags:
      - employees
  /employees/{id}/ratings:
    get:
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/employees.RatingItem'
                  type: array
              type: object
      summary: List ratings given to a master (as an employee)
      tags:
      - employees
    post:
      consumes:
      - application/json
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: string
      - description: User + score
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/employees.RateEmployeeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/employees.RatingAgg'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.Envelope'
      summary: Rate a master (1..5)
      tags:
      - employees
  /employees/{id}/recipe-feedback:
    get:
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/employees.RecipeFeedbackItem'
                  type: array
              type: object
      summary: Reviews & scores on a master's recipes (with components)
      tags:
      - employees
  /employees/{id}/restaurants:
    post:
      consumes:
      - application/json
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: string
      - description: Restaurant code
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/employees.AttachRestaurantRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.Envelope'
      summary: Attach an employee to another restaurant (by code)
      tags:
      - employees
  /employees/batch:
    post:
      consumes:
      - application/json
      parameters:
      - description: Employee ids
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/employees.BatchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/employees.EmployeeResponse'
                  type: array
              type: object
      summary: Get employees by ids (batch)
      tags:
      - employees
...
  /restaurants:
    post:
      consumes:
      - application/json
      parameters:
      - description: Restaurant name
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/employees.CreateRestaurantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/httpx.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/employees.RestaurantResponse'
              type: object
      summary: Create a restaurant and get its code
      tags:
      - restaurants
  /restaurants/{restaurantId}/shift:
    get:
      parameters:
      - description: Restaurant ID
        in: path
        name: restaurantId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/employees.ShiftMaster'
                  type: array
              type: object
      summary: Today's masters on shift in a restaurant
      tags:
      - restaurants
  /restaurants/employees:
    post:
      consumes:
      - application/json
      parameters:
      - description: Restaurant code
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/employees.ListByCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/employees.LoginResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.Envelope'
      summary: List restaurant employees by code (for login)
      tags:
      - restaurants
  /restaurants/shift:
    post:
      consumes:
      - application/json
      parameters:
      - description: Code + employee ids
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/employees.SetShiftRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/employees.ShiftMaster'
                  type: array
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.Envelope'
      summary: Set today's masters on shift (by restaurant code)
      tags:
      - restaurants
...
swagger: "2.0"